+++ CHANGELOG.md
@@ -7,6 +7,8 @@
   * [#301](https://github.com/mozilla-rally/core-addon/pull/301): Correctly report the zip code in the demographics survey.
   * [#298](https://github.com/mozilla-rally/core-addon/pull/298): Disable Rally on locales other than `en-US`. The ` --config-disable-locale-check` build option allows overriding the check for developer workflows on other locales.
   * [#305](https://github.com/mozilla-rally/core-addon/pull/305): Disable data submission to enable safer QA.
+* `rally.js`
+  * [#306](https://github.com/mozilla-rally/core-addon/pull/306): Add a developer mode. When enabled it allows developer to dump the content of pings to the browser console and prevents the Rally information page to be opened if no core-addon is found.
 
 # v0.7.1 (2021-01-13)